Area Loader
Added an area loader which will load the next area when a trigger is
entered.  Also some modifications to the lightning to take into account
the players position when deciding if it should respawn at - the axis
its reached the maximum, and the heartbeat script now takes into account
the volume.