HUD and Camera Fade
Added in a HUD element to indicate an item is able to be picked up, and
also a camera color fade effect.  Also added in some new controls, C
will turn the cursor off and on so its easier to walk around and select
things, and F will trigger a camera fade in.