Item Holders
Items can now be placed on a shelf i have put on the container.  There
are three slots on the shelf to put them on, a different icon to the
hand pickup icon will appear when you are aiming at it, then click the
button on the side of the hand your item is, and it should go on there.