Autosave
Added an autosave feature to Unity.  Click on the Autosave option in the
menu at the top to open up a little window to turn it off and on and set
the amount of time there is between autosaves.  The window is idealy
suited to be docked underneath the hierachy window.