Some stuff
Tidied up the dialogue stuff, rather than putting a copy of a dialogue
into the trigger it now just puts the key there, which is then used to
access the dialogue data from a dialogue holder, which can now be used
as the game is running.  Also, added a Thing. Look out the window while
your in the house.